Show alert and hide the cart after submit.

+++ src/components/ShoppingCart/ShoppingCart.js
@@ -8,31 +8,39 @@ import MainCard from './MainCard/MainCard';
 import ShoppingSummary from './ShoppingSummary/ShoppingSummary';
 
 const ShoppingCart = (props) => {
-    return(
-        <div className={styles.ShoppingCart}>
-            <div className={styles.HeaderContainer}>
-                <div className={styles.Title}>
-                    {props.header.title}
+    const ShoppingCart = () => {
+        if (!props.isVisible){
+            return;
+        }
+        return (
+            <div className={styles.ShoppingCart}>
+                <div className={styles.HeaderContainer}>
+                    <div className={styles.Title}>
+                        {props.header.title}
+                    </div>
+                    <Button 
+                        text={props.header.buttonText}
+                        onClickFunction={props.orderSubmited}
+                    />
                 </div>
-                <Button 
-                    text={props.header.buttonText} 
-                />
+                <div className={styles.MainContainer}>
+                    <MainCard 
+                        mainCardText={props.mainCardText}
+                        items={props.items}
+                        setQuantity={props.setQuantity}
+                        updateTotals={props.updateTotals}
+                        deleteItem={props.deleteItem}
+                    />
+                    <ShoppingSummary 
+                        shoppingSummaryText={props.shoppingSummaryText}
+                        totals={props.totals}
+                        orderSubmited={props.orderSubmited}
+                    />
+                </div>
             </div>
-            <div className={styles.MainContainer}>
-                <MainCard 
-                    mainCardText={props.mainCardText}
-                    items={props.items}
-                    setQuantity={props.setQuantity}
-                    updateTotals={props.updateTotals}
-                    deleteItem={props.deleteItem}
-                />
-                <ShoppingSummary 
-                    shoppingSummaryText={props.shoppingSummaryText}
-                    totals={props.totals}
-                />
-            </div>
-        </div>
-    );
+        );
+    }
+    return ShoppingCart();
 }
 
 export default ShoppingCart;
